backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Pinterest AI Agent Backend"
authors = [{ name = "shniss", email = "seanhnissenbaum@gmail.com" }]
readme = "README.md"

# allow any recent 3.x you're likely to use locally or in Docker
requires-python = ">=3.11,<3.14"

dependencies = [
    # Core FastAPI and async
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn[standard] (>=0.35.0,<0.36.0)",
    "websockets (>=12.0,<13.0)",
    
    # Database
    "motor (>=3.7.1,<4.0.0)",
    "pymongo (>=4.7.0,<5.0.0)",
    
    # Background tasks
    "celery[redis] (>=5.5.3,<6.0.0)",
    "redis (>=5.0.0,<6.0.0)",
    
    # Web scraping and automation
    "playwright (>=1.40.0,<2.0.0)",
    "beautifulsoup4 (>=4.12.0,<5.0.0)",
    "requests (>=2.31.0,<3.0.0)",
    
    # AI and validation
    "openai (>=1.0.0,<2.0.0)",
    "pillow (>=10.0.0,<11.0.0)",
    
    # Utilities
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pydantic[email] (>=2.11.7,<3.0.0)",
    "httpx (>=0.25.0,<1.0.0)",
    "aiofiles (>=23.0.0,<24.0.0)",
    "nest-asyncio (>=1.5.0,<2.0.0)",
    "spacy (>=3.8.7,<4.0.0)",
    "black (>=25.1.0,<26.0.0)",
    "ruff (>=0.12.7,<0.13.0)"
]

[tool.poetry]
packages = [
    { include = "app" }
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
